# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraBaseBD Reseller Place API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  stock: number;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  createdAt: string;
}