# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraBaseBD Reseller Place API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  price: string;
  stock: number;
  imageUrl?: string;
}