# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AuraBaseBD Reseller Place API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  name?: string;
  email?: string;
  password?: string;
  role?: string;
  phone?: string;
  isActive?: boolean;
}